Simplify and improve reuse of classes in hmc module.
All MC classes used by the interface in trimem/mc/util now access
the mesh through the EnergyManager to have a more consistent view
on the state of the mesh.

The MeshFlips class was redesigned to be used for several mesh
mutation operations. This is in preparation for the upcoming
support of swapping of vertex properties.

+++ src/flips.h
@@ -12,8 +12,8 @@ namespace trimem {
 
 struct EnergyManager;
 
-int flip_serial(EnergyManager& estore, const real& flip_ratio);
+double flip_serial(EnergyManager& estore, const real& flip_ratio);
 
-int flip_parallel_batches(EnergyManager& estore, const real& flip_ratio);
+double flip_parallel_batches(EnergyManager& estore, const real& flip_ratio);
 }
 #endif
